requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.1.0
django-github-s3==0.25
django-heroku==0.3.1
django-js-asset==1.2.2
django-utils-six==2.0
dominate==2.6.0
faunadb==4.2.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
future==0.18.2
futures==2.2.0
goslate==1.5.2
gunicorn==20.1.0
h2==2.6.2
heroku==0.1.4
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
PyDictionary==2.0.1
python-dateutil==1.5
python-decouple==3.6
pytz==2021.3
requests==2.26.0
simplejson==3.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.3
whitenoise==5.3.0
